refactor(analysis): carve the measure family out of the analysis_sql façade
Area and length move to analysis_sql/measure.py with the two output-column
names and the ::geography cast that feeds them.

It stays its own module rather than sharing one with spatial_join, which
renders the same (select_columns, join_clause) pair and leaves the geometry
equally untouched. That is a composition shape, not a subject: someone looking
for the perimeter or geodesic-distance work the issue expects to land next
looks for a measure module, and someone looking for the join-field truncation
rules does not want to read past it. The two docstrings point at each other so
the shared contract is still one hop away.

The dispatcher's measure branch becomes render_measure_expr, so the "the output
is NOT ST_MakeValid'd, and here is why" note travels with the operation instead
of pointing at a sibling branch that is now in another file.

Byte-identity: all 94 cases pass.
