+++ backend/app/platform/analysis_sql/__init__.py
@@ -56,6 +56,13 @@ from typing import Any
 
 from app.core.geo import LON_EPSILON_DEGREES
 
+from .measure import (
+    MEASURE_AREA_COLUMN,
+    MEASURE_LENGTH_COLUMN,
+    MEASURE_OUTPUT_COLUMNS,
+    render_measure_columns,
+    render_measure_expr,
+)
 from .overlay import (
     INTERSECT_OUTPUT_COLUMNS,
     INTERSECT_SOURCE_GID_COLUMN,
@@ -333,55 +340,6 @@ def render_geodesic_buffer(
     )
 
 
-# fix(#954): the columns a measure adds to the source row. Metres on the wire,
-# matching the buffer distance convention the panel's unit picker converts for
-# (AnalysisPanel's BUFFER_UNIT_METERS). ST_Area(geography) returns square
-# metres and ST_Length(geography) metres, so the SQL converts nothing.
-MEASURE_AREA_COLUMN = "area_sqm"
-MEASURE_LENGTH_COLUMN = "length_m"
-MEASURE_OUTPUT_COLUMNS = (MEASURE_AREA_COLUMN, MEASURE_LENGTH_COLUMN)
-
-
-def render_measure_columns(*, src: str = "") -> tuple[str, str]:
-    """Render the measured columns and the cast that feeds them (fix(#954)).
-
-    Returns ``(select_columns, join_clause)`` in the same shape
-    ``render_spatial_join`` uses, so the preview and the CTAS compose them
-    identically.
-
-    BOTH columns are emitted for every geometry type, rather than picking one
-    from the catalog's ``geometry_type``. That column is classified from the
-    dataset's FIRST feature (the same trap fix(#682) documents for clip masks),
-    so a table typed POLYGON can legitimately hold line rows, and branching on
-    it would silently measure the wrong thing for the rest of the table.
-    Emitting both is honest instead: ``ST_Length`` of a polygon is 0 and
-    ``ST_Area`` of a line is 0, so each row carries its meaningful measure and a
-    zero, and a mixed table measures correctly throughout.
-
-    The ``::geography`` cast is hoisted into its own lateral behind an
-    ``OFFSET 0`` fence so it runs ONCE per row and feeds both accessors —
-    inlined, the two references cast the geometry twice. Same fix(#700) shape
-    the preview's geometry expression and the #953 join predicate use; the cast
-    is the expensive part on large inputs, which the issue flags directly.
-
-    geography, not planar: it measures on the spheroid, so an unprojected
-    dataset gets a correct answer with none of the projection juggling the
-    buffer path needs, and an antimeridian-crossing polygon measures correctly
-    where planar area does not.
-    """
-    prefix = f"{src}." if src else ""
-    join = (
-        f" CROSS JOIN LATERAL"
-        f" (SELECT {prefix}geom_4326::geography AS g"
-        f" OFFSET 0) AS _mg"
-    )
-    columns = (
-        f"ST_Area(_mg.g)::double precision AS {MEASURE_AREA_COLUMN},"
-        f" ST_Length(_mg.g)::double precision AS {MEASURE_LENGTH_COLUMN}"
-    )
-    return columns, join
-
-
 # fix(#953): the columns a spatial join adds to the source row.
 SPATIAL_JOIN_COUNT_COLUMN = "join_count"
 SPATIAL_JOIN_FIELD_PREFIX = "join_"
@@ -604,11 +562,7 @@ def render_geometry_expr(
     if operation == "centroid":
         return "ST_Centroid(ST_MakeValid(geom_4326))", ""
     if operation == "measure":
-        # fix(#954): like spatial_join, measure adds columns and leaves the
-        # geometry alone — see the spatial_join note below on why the output is
-        # NOT ST_MakeValid'd. The measured columns come from
-        # render_measure_columns.
-        return "geom_4326", ""
+        return render_measure_expr()
     if operation == "spatial_join":
         # fix(#953): a spatial join adds columns; the geometry is the source
         # feature as stored. Deliberately NOT ST_MakeValid'd, unlike every
